[MOV] base: move create menu wizard in its own file
Does not really belong to ir.model.
