+++ odoo/addons/base/models/ir_model.py
@@ -2642,27 +2642,3 @@ class IrModelData(models.Model):
         self.env[model].browse(res_id).check_access('write')
         for xid in self.search([('model', '=', model), ('res_id', '=', res_id)]):
             xid.noupdate = not xid.noupdate
-
-
-class WizardIrModelMenuCreate(models.TransientModel):
-    _name = 'wizard.ir.model.menu.create'
-    _description = 'Create Menu Wizard'
-
-    menu_id = fields.Many2one('ir.ui.menu', string='Parent Menu', required=True, ondelete='cascade')
-    name = fields.Char(string='Menu Name', required=True)
-
-    def menu_create(self):
-        for menu in self:
-            model = self.env['ir.model'].browse(self._context.get('model_id'))
-            vals = {
-                'name': menu.name,
-                'res_model': model.model,
-                'view_mode': 'list,form',
-            }
-            action_id = self.env['ir.actions.act_window'].create(vals)
-            self.env['ir.ui.menu'].create({
-                'name': menu.name,
-                'parent_id': menu.menu_id.id,
-                'action': 'ir.actions.act_window,%d' % (action_id,)
-            })
-        return {'type': 'ir.actions.act_window_close'}
