+++ odoo/addons/base/wizard/__init__.py
@@ -8,3 +8,4 @@ from . import base_module_upgrade
 from . import base_module_uninstall
 from . import base_export_language
 from . import base_partner_merge
+from . import wizard_ir_model_menu_create
