+++ odoo/addons/base/wizard/wizard_ir_model_menu_create.py
@@ -1,0 +1,25 @@
+from odoo import fields, models
+
+
+class WizardIrModelMenuCreate(models.TransientModel):
+    _name = 'wizard.ir.model.menu.create'
+    _description = 'Create Menu Wizard'
+
+    menu_id = fields.Many2one('ir.ui.menu', string='Parent Menu', required=True, ondelete='cascade')
+    name = fields.Char(string='Menu Name', required=True)
+
+    def menu_create(self):
+        for menu in self:
+            model = self.env['ir.model'].browse(self._context.get('model_id'))
+            vals = {
+                'name': menu.name,
+                'res_model': model.model,
+                'view_mode': 'list,form',
+            }
+            action_id = self.env['ir.actions.act_window'].create(vals)
+            self.env['ir.ui.menu'].create({
+                'name': menu.name,
+                'parent_id': menu.menu_id.id,
+                'action': 'ir.actions.act_window,%d' % (action_id,)
+            })
+        return {'type': 'ir.actions.act_window_close'}
